plugins/elements/junction.py: Remove ambiguous cache related junction parameters
This removes the "cache-junction-elements" and "ignore-junction-remotes"
parameters from the junction YAML configuration.

As referenced by the mailing list thread[0], these parameters do not belong
in element configuration, fine grained configuration of these parameters
are not needed in the context of a pipeline with many subprojects in play,
and the documentation and expected behavior with deep and complex subproject
hierarches is unclear, especially when considering the ability to override
junction configurations. Further, these currently cannot even be overridden
by user configuration.

This will be replaced by a simpler configuration in the regular way, with
recommendations (not hard decisions) in project.conf and the final decision
made by user configuration files.

Summary of changes:

  * plugins/elements/junction.py: Removed the configuration and docs

  * _project.py: Stop observing these configurations

  * tests/artifactcache/junctions.py: Stop testing these configurations

[0]: https://lists.apache.org/thread.html/rf2da9830e2fa918357f99a6021e55fc43df876f0b19d43f68802f083%40%3Cdev.buildstream.apache.org%3E