Refactor remote asset user configuration
This is a full rewrite of how configuration data pertaining to
AssetCaches, such as the artifact cache and source caches is loaded, which
before this refactor used to be spread out across several modules.

With the previous refactor of RemoteSpec/RemoteExecutionSpec, we now
have the configuration loading performed by the appropriate modules
Context and Project, where the final decision is made by Context, at
which point the Context will dictate to the AssetCaches which remotes
to use in which way.

In addition to this refactor, the priorities of remotes are brought more
in line with BuildStream general policy for user configuration data, so
instead of compounding remotes found at different levels, we allow user
configuration to properly override remotes suggested by project configurations,
the order is now:

  * CLI provided `--remote` arguments override anything else, for all projects
    being processed in the given session.

  * Remotes defined in buildstream.conf (user configuration) on a per project
    basis specify the full list of remotes which will be used to process that project,
    overriding anything at lower levels.

  * Globally defined remotes in the user configuration provide the defaults
    for processing all projects in the session, overriding any suggestions
    made by project defined recommendations.

  * Remotes defined in project.conf (project configuration) are treated
    as recommendations, and are overridden by afore mentioned user configuration.

This priority shuffle is inline with the proposal[0]

Another side effect of this refactor is that we no longer support specifying
a single asset cache for any of the cache types, they must always be
specified in YAML as lists.

Rationale: It is more clear that it is a list if we force it to be a list.
This feature could have been useful for convenience, except that it adds no
convenience at all; when specifying a list with a single element or just
specifying the dictionary directly, there is no additional yaml (only one space
becomes a '-' instead of a ' ').

Seeing as it wasn't buying any convenience, it seems pointless to keep
this around.

Summary of changes:

  * _project.py:

    Now parse the project recommended remotes using RemoteSpec.new_from_node()

  * _context.py:

    Now parse the global remotes and project specific overrides using RemoteSpec.new_from_node()

    Also Context is now in charge of initializing the asset caches, which
    already technically belong to the Context object, which provides the
    properties which are responsible for lazy instantiation of the caches.

    A new Context.initialize_remotes() method is added here for that purpose.

  * _stream.py:

    Updated to call Context.initialize_remotes() instead of delegating this activity
    to the AssetCaches themselves.

    Also renamed the `use_artifact_config` and `use_source_config` arguments used in
    this file to `connect_artifact_cache` and `connect_source_cache`, this simplifies
    the code slightly.

  * _assetcache.py:

    Any remaining parsing code has been removed from here.

    All of the code which used to decide which cache to use in which order
    has been replaced, an asset cache is now given the full list of
    RemoteSpec objects it is expected to communicate with in a given session,
    accompanied by a dictionary of RemoteSpec lists which define which remotes
    to use for each given project it might encounter during the session.

    A new method AssetCache.get_remotes() is added for the convenience of
    subclasses so that they might obtain the remotes they need.

  * _artifactcache.py, _sourcecache.py, _elementsourcescache.py:

    Updated for new API changes, mostly this means using AssetCache.get_remotes().

  * tests: Updated test cases to ensure we always specify caches as lists.

[0]: https://lists.apache.org/thread.html/rf2da9830e2fa918357f99a6021e55fc43df876f0b19d43f68802f083%40%3Cdev.buildstream.apache.org%3E